Problem: some CZMQ methods are using fractured return values
Mainly methods like zmsg_encode and zarmour_encode, which return
data via an argument and size via the return code. This is nasty
for many reasons. It makes it impossible to create clean bindings
on top, for one.

Solution: ban this behavior. If a method returns a size-delimited
buffer it should use an appropriate object (e.g. zchunk).